Move API services to application layer
I'm going to make a new API system in the future, so for now these
services will be made for internal use cases in its own layer, following
the clean architecture more stringently.
