+++ src/main/kotlin/dev/mizarc/waystonewarps/application/PlayerUnregisterResult.kt
@@ -1,4 +1,4 @@
-package dev.mizarc.waystonewarps.api.enums
+package dev.mizarc.waystonewarps.application.enums
 
 /**
  * Represents the result of unregistering a player's state.
